+++ inst/shinyDW/global.R
@@ -1,0 +1,4 @@
+suppressPackageStartupMessages(library(dwtools))
+library(shinyTree)
+
+numcols <- sapply(x,class)=="numeric"